requirements.txt
streamlit 
torchvision
pillow

